Remove storage.rs. Move Indexed to attestation.rs
