+++ types/src/consensus/attestation.rs
@@ -1,3 +1,4 @@
+use alloy_sol_types::SolValue;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 
 use crate::{
@@ -7,7 +8,6 @@ use crate::{
         signer::UncheckedSigned,
     },
     ledger::receipt::UncheckedReceipt,
-    storage::Indexed,
     Receipt, Signed, Timestamp, Transaction,
 };
 
@@ -17,6 +17,40 @@ pub type ReceiptAttestation = Attestation<Receipt>;
 pub type UncheckedReceiptAttestation = Attestation<UncheckedReceipt>;
 
 pub type UncheckedTransactionAttestation = Attestation<UncheckedSigned<Transaction>>;
+
+#[derive(Clone, Serialize, Deserialize, Debug)]
+pub struct Indexed<T> {
+    #[serde(rename = "timestamp")]
+    pub index: Timestamp, // TODO: consider sequential numbers
+    pub value: T,
+}
+
+impl<T> Indexed<T> {
+    pub fn new(index: Timestamp, value: T) -> Self {
+        Indexed { index, value }
+    }
+}
+
+impl<T> From<Indexed<Attestation<T>>> for Indexed<HeadlessAttestation> {
+    fn from(value: Indexed<Attestation<T>>) -> Self {
+        Indexed {
+            index: value.index,
+            value: value.value.into(),
+        }
+    }
+}
+
+impl<T: Hashable> Hashable for Indexed<T> {
+    fn hash_custom(&self) -> Hash {
+        alloy_primitives::keccak256(
+            [
+                self.index.as_micros().abi_encode(),
+                self.value.hash_custom().to_vec(),
+            ]
+            .concat(),
+        )
+    }
+}
 
 // An Attestation<T> is T signed by a validator using ECDSA
 #[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
